dev: remove pretty stuff
