+++ doc/notes/2025-08-23-retro.md
@@ -1,0 +1,64 @@
+# 2025-08-23 Retro: Renderer and Logging Rework
+
+## Summary
+Over the `pretty-format-improvements` branch, we split rendering concerns and redesigned logging. This doc captures the decisions so we can restart cleanly while keeping the learnings.
+
+## Renderer Architecture
+- Alt-screen renderer for ratatui UI
+  - New: `src/prompt/renderer/alt_screen.rs`
+  - Draws logs and prompt within the terminal alternate screen frame.
+- Stdout renderer for regular terminal output
+  - New: `src/prompt/renderer/stdout.rs`
+  - Prints logs to terminal scrollback and keeps the prompt at the bottom.
+- Shared helpers and config
+  - Updated: `src/prompt/renderer/helpers.rs`
+  - New: `src/prompt/config.rs`
+- Removed older renderers
+  - Removed: `src/prompt/renderer/pretty.rs`, `src/prompt/renderer/minimal.rs`
+- Integration points
+  - Updated: `src/prompt/mod.rs`, `src/prompt/renderer/mod.rs`
+
+## Logging Strategy
+Two strategies depending on renderer:
+- Stdout strategy (for stdout renderer)
+  - Capture subscriber + `repl_println!` prints
+  - Logs go to terminal scrollback; prompt remains visible at bottom
+- In-frame buffer (for alt-screen/ratatui)
+  - `LogBufferPlugin` collects tracing events into a buffer
+  - Renderer draws logs inside the ratatui frame above the prompt block
+- Important: Do not enable both strategies at once to avoid duplicate logs. Choose via `PromptRenderer::strategy()` (renderer decides the logging display).
+- Relevant code: `src/log_ecs.rs`, `src/print.rs`, renderer modules above
+
+## Plugins, Features, and Wiring
+- `src/plugin.rs`: respect feature flags; `PromptPlugin::default` wires in renderers/logging based on enabled features
+- `src/lib.rs`, `src/repl.rs`: simplified exports and REPL setup consistent with renderer split
+- `src/command/parser.rs`, `src/built_ins/clear.rs`: small adjustments
+
+## Examples and Demos
+- New/updated examples to showcase both flows:
+  - `examples/simple.rs`, `examples/simple_sandbox.rs`, `examples/pretty_sandbox.rs`
+  - `examples/custom_renderer.rs`, `examples/pretty.rs`, `examples/pretty_custom.rs`
+  - `examples/log_custom_layer.rs` (and removed `examples/log_minimal.rs`)
+
+## Tests
+- Added tests and support:
+  - `tests/tui_prompt_widget.rs` with snapshot at `tests/snapshots/tui_prompt_widget__pretty_prompt_snapshot.snap.new`
+  - `tests/renderer_adapters.rs`, `tests/e2e_pretty_pty.rs`
+  - `src/test_support/mod.rs`
+
+## Docs & README
+- README significantly updated for new renderer/logging model
+- Design notes:
+  - `doc/notes/2025-08-16-repl-logging-design.md`
+  - `doc/notes/2025-08-17-logging-strategies.md`
+  - `doc/notes/2025-08-19-testing-strategy.md`
+
+## What Worked / What Didn’t
+- Worked: explicit split between alt-screen vs stdout renderers; clearer mental model; easier testing.
+- Mixed: multiple renderers in parallel caused duplication; picking one strategy per renderer is key.
+- Next: firm up adapter interfaces and reduce surface in `src/prompt/mod.rs`.
+
+## Next Steps After Reset
+- Keep renderer split and logging strategy as the baseline design.
+- Reintroduce minimal, well-scoped features incrementally with tests.
+- Verify defaults via examples `examples/simple.rs` and `examples/pretty.rs`.
