+++ examples/log.rs
@@ -6,9 +6,9 @@ struct SpamTimer(Timer);
 
 fn setup(mut commands: Commands) {
     commands.insert_resource(SpamTimer(Timer::from_seconds(0.5, TimerMode::Repeating)));
-    bevy_repl::repl_println!("\nBevy log routing example");
+    bevy_repl::repl_println!("\nBevy log routing example (MinimalRenderer)");
     bevy_repl::repl_println!(
-        "Logs emitted by Bevy/tracing are captured and printed above the REPL prompt."
+        "Logs emitted by Bevy/tracing are printed above the prompt using the minimal renderer."
     );
     bevy_repl::repl_println!("Type `quit` to exit.");
 }
@@ -25,18 +25,16 @@ fn spam_logs(mut timer: ResMut<SpamTimer>, time: Res<Time>) {
 
 fn main() {
     // Install a global fmt layer that writes logs directly to the REPL printer,
-    // preserving colors and formatting. Do this BEFORE adding DefaultPlugins.
-    tracing_to_repl_fmt_with_level(bevy::log::Level::DEBUG);
+    // preserving colors/formatting. Do this BEFORE adding DefaultPlugins.
+    tracing_to_repl_fmt();
 
     App::new()
         .add_plugins((
-            // Disable stdout logger to avoid duplicate output; our fmt layer handles printing
+            // Disable stdout logger to avoid duplicate output; our fmt layer prints
             DefaultPlugins.build().disable::<bevy::log::LogPlugin>(),
-            // Enable the REPL (pretty recommended for the prompt scroll region)
-            ReplPlugins.set(PromptPlugin::pretty()),
+            ReplPlugins,
         ))
         .add_systems(Startup, setup)
-        // Emit logs at various levels
         .add_systems(Update, spam_logs)
         .run();
 }
