--- examples/pretty_custom.rs
@@ -1,65 +1,0 @@
-//! Pretty REPL prompt example (Manual composition).
-//!
-//! This example shows configuring the pretty styling by composing the
-//! plugins manually and passing a custom `PromptPlugin` with a
-//! `ReplPromptConfig` preset.
-//!
-//! Run with: `cargo run --example pretty_manual --features pretty`
-
-use bevy::{app::ScheduleRunnerPlugin, prelude::*};
-use bevy_repl::{
-    prelude::*,
-    prompt::{
-        PromptBorderConfig, PromptColorConfig, PromptHintConfig, renderer::pretty::PrettyRenderer,
-    },
-};
-use std::sync::Arc;
-use std::time::Duration;
-
-#[derive(Debug, Clone, Event, Default)]
-struct PingCommand;
-
-impl ReplCommand for PingCommand {
-    fn clap_command() -> clap::Command {
-        clap::Command::new("ping").about("Test command")
-    }
-}
-
-fn on_ping(_trigger: Trigger<PingCommand>) {
-    repl_println!("Pong");
-}
-
-fn instructions() {
-    repl_println!();
-    repl_println!("Welcome to the Bevy REPL pretty prompt example (manual)!");
-    repl_println!();
-    repl_println!("Try typing a command:");
-    repl_println!("  `ping`    - Trigger the ping command. (it outputs Pong)");
-    repl_println!("  `quit`    - Close the app.");
-    repl_println!();
-    repl_println!("Press CTRL+C to exit any time.");
-    repl_println!();
-}
-
-fn main() {
-    App::new()
-        .add_plugins((
-            MinimalPlugins.set(ScheduleRunnerPlugin::run_loop(Duration::from_secs_f64(
-                1.0 / 60.0,
-            ))),
-            bevy::input::InputPlugin::default(),
-            ReplPlugins.set(PromptPlugin {
-                config: ReplPromptConfig {
-                    symbol: Some("λ ".into()),
-                    border: Some(PromptBorderConfig::default()),
-                    color: Some(PromptColorConfig::default()),
-                    hint: Some(PromptHintConfig::default()),
-                },
-                renderer: Arc::new(PrettyRenderer),
-            }),
-        ))
-        .add_repl_command::<PingCommand>()
-        .add_observer(on_ping)
-        .add_systems(PostStartup, instructions.after(ScrollRegionReadySet))
-        .run();
-}
