+++ src/prompt/renderer/mod.rs
@@ -1,7 +1,6 @@
 pub mod minimal;
 pub mod helpers;
-#[cfg(feature = "pretty")]
-pub mod pretty;
+pub mod scroll;
 
 use bevy::prelude::*;
 use bevy_ratatui::RatatuiContext;
@@ -12,9 +11,6 @@ use ratatui::layout::Rect;
 use std::sync::Arc;
 
 /// Public label: "scroll region ready". Always available, even in minimal mode.
-#[derive(SystemSet, Debug, Hash, PartialEq, Eq, Clone)]
-pub struct ScrollRegionReadySet;
-
 pub struct PromptRenderPlugin {
     pub renderer: Arc<dyn PromptRenderer>,
 }
@@ -40,8 +36,6 @@ pub struct ActiveRenderer(pub Arc<dyn PromptRenderer>);
 impl Plugin for PromptRenderPlugin {
     fn build(&self, app: &mut App) {
         app.insert_resource(ActiveRenderer(self.renderer.clone()));
-        // Expose the PostStartup ready set unconditionally so callers can order after it.
-        app.configure_sets(PostStartup, ScrollRegionReadySet);
         app.add_systems(
             Update,
             (
@@ -52,8 +46,9 @@ impl Plugin for PromptRenderPlugin {
                     .after(ReplSet::Buffer),
             ),
         );
-        #[cfg(feature = "pretty")]
-        app.add_plugins(pretty::ScrollRegionPlugin);
+        app.add_plugins(
+            scroll::ScrollRegionPlugin,
+        );
     }
 }
